RaspberryPiCode/requirements.txt
pyserial
paramiko
opencv-python
numpy